Habit checkboxes & input
