+++ components/DashboardPage/components/HabitTracker/Habit.module.scss
@@ -1,0 +1,15 @@
+.Text {
+  appearance: none;
+  -webkit-appearance: none;
+  background-color: transparent;
+  border: none;
+  font-weight: 400;
+  font-size: 12px;
+  color: var(--secondaryTextColor);
+  width: 100%;
+  padding-left: 8px;
+}
+
+.Text:focus {
+  outline: none;
+}
