+++ components/DashboardPage/components/HabitTracker/Habit.tsx
@@ -1,4 +1,6 @@
 import { Stack, Typography } from 'components';
+import { Checkbox } from './Checkbox';
+import styles from './Habit.module.scss';
 
 interface HabitProps {
   label: string;
@@ -8,9 +10,23 @@ export function Habit({ label }: HabitProps) {
   return (
     <Stack width="100%" spacing="tight">
       <Stack width="30%">
-        <Typography type="mediumText">{label}</Typography>
+        <input
+          type="text"
+          placeholder="Add a habit"
+          className={styles.Text}
+        />
       </Stack>
-      <Stack width="70%">ok</Stack>
+      <Stack width="70%">
+        <Stack width="100%" spread>
+          <Checkbox />
+          <Checkbox />
+          <Checkbox />
+          <Checkbox />
+          <Checkbox />
+          <Checkbox />
+          <Checkbox />
+        </Stack>
+      </Stack>
     </Stack>
   );
 }
